Refactored classes
