+++ Package/Editor/CrashKonijn.Goap.Editor/GraphViewer/ConditionElement.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 5e8f656a20d840769929724a0dde32bf
+timeCreated: 1712683465
