+++ Package/Editor/CrashKonijn.Goap.Editor/GraphViewer/ConnectionElement.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 9b579596134442708e41b66b6e9803b6
+timeCreated: 1712683465
