+++ Package/Editor/CrashKonijn.Goap.Editor/GraphViewer/EffectElement.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 3bb9750c31f34631a4c312c4e85efdcf
+timeCreated: 1712683465
