+++ Package/Editor/CrashKonijn.Goap.Editor/GraphViewer/NodeElement.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 0884da2127404057bc3dfe2c6dcd91ad
+timeCreated: 1712683465
