+++ Package/Editor/CrashKonijn.Goap.Editor/GraphViewer/SelectedObject.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: d18597f0026f4624a32f9881a159b6ae
+timeCreated: 1712683465
