+++ Package/Editor/CrashKonijn.Goap.Editor/Styles/GraphViewer.uss
