--- Package/Editor/CrashKonijn.Goap.Editor/Styles/Test_2022.uss
@@ -1,10 +1,0 @@
-﻿/*.drag-root {*/
-/*    background-color: var(--unity-colors-alternated_rows-background);*/
-/*    background-image: url('project://database/Packages/com.crashkonijn.goap/Editor/CrashKonijn.Goap.Editor/Textures/background_grid.png');*/
-/*    background-position: 0 0;*/
-/*    background-repeat: repeat;*/
-/*    background-size: 100px 100px;*/
-/*    */
-/*    !*border-width: 1px;*!*/
-/*    !*border-color: red;*!*/
-/*}*/
