--- Package/Editor/CrashKonijn.Goap.Editor/Styles/Test_2022.uss.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 98bdb9a43d814927a148628c65292000
-timeCreated: 1712606395
