--- Package/Editor/CrashKonijn.Goap.Editor/Test.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d76340f9ce5705b4bb93997747cac44f
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
